requirements.txt
beautifulsoup4
discord.py
mysql-connector-python
python-dotenv
requests
